scraper/requirements.txt
requests-html
grpcio-tools
asyncio
fake-useragent
redis